requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
cryptography==2.9
cycler==0.10.0
cytoolz==0.10.1
dask==2.14.0
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
idna==2.9
imageio==2.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.4
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
Pillow==7.0.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
soupsieve==2.0
toolz==0.10.0
tornado==6.0.4
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
